web-animations/animation-player/animation-player-play.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Animation player play() method test</title>
<meta name="assert" content="Unpauses the player and rewinds if it has finished playing">
<link rel="help" href="http://dev.w3.org/fxtf/web-animations/#widl-AnimationPlayer-play-void">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
async_test(function(t) {
    var player = document.timeline.play(null);

    player.ready.then(t.step_func(function() {
        player.pause();
        player.play();

        assert_equals(player.playState, 'finished', 'AnimationPlayer play state ' +
            'should be finished');
        t.done();
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'AnimationPlayer.play() unpauses the player even if player has no source content');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(t));
    var timesToTest = [
        -100,
        -1,
        0,
        1,
        100,
        ANIMATION_END_TIME / 2,
        ANIMATION_END_TIME - 100,
        ANIMATION_END_TIME - 1,
        ANIMATION_END_TIME,
        ANIMATION_END_TIME + 1,
        ANIMATION_END_TIME + 100];

    player.ready.then(t.step_func(function() {
        timesToTest.forEach(function(time) {
            player.currentTime = time;
            player.pause();
            player.play();

            assert_not_equals(player.playState, 'paused', 'AnimationPlayer should be ' +
                'unpaused for time ' + time);
        });
        t.done();
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'AnimationPlayer.play() unpauses the player. Test different times');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(t));
    player.playbackRate = -1;
    var timesToTest = [
        -100,
        -1,
        0,
        1,
        100,
        ANIMATION_END_TIME / 2,
        ANIMATION_END_TIME - 100,
        ANIMATION_END_TIME - 1,
        ANIMATION_END_TIME,
        ANIMATION_END_TIME + 1,
        ANIMATION_END_TIME + 100];

    player.ready.then(t.step_func(function() {
        timesToTest.forEach(function(time) {
            player.currentTime = time;
            player.pause();
            player.play();

            assert_not_equals(player.playState, 'paused', 'AnimationPlayer should be ' +
                'unpaused for time ' + time);
        });
        t.done();
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'AnimationPlayer.play() unpauses the player when playback rate < 0');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));

    player.ready.then(t.step_func(function() {
        // wait for 50 ms
        setTimeout(function() {
            t.step(function () {
                player.pause();
                // wait for 50 ms
                setTimeout(function() {
                    t.step(function() {
                        player.play();
                        // wait for 50 ms
                        setTimeout(function() {
                            t.step(function() {
                                assert_approx_equals(player.currentTime, 100, EPSILON,
                                    'After player is unpaused by AnimationPlayer.play() ' +
                                    'current time should go on increasing');
                                t.done();
                            });
                        }, 50);
                    });
                }, 50);
            });
        }, 50);
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'Test that after player is unpaused by AnimationPlayer.play() ' +
    'current time goes on increasing');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));
    player.playbackRate = -1;
    player.currentTime = ANIMATION_END_TIME / 2;

    player.ready.then(t.step_func(function() {
        // wait for 50 ms
        setTimeout(function() {
            t.step(function () {
                player.pause();
                // wait for 50 ms
                setTimeout(function() {
                    t.step(function() {
                        player.play();
                        // wait for 50 ms
                        setTimeout(function() {
                            t.step(function() {
                                var expectedTime = ANIMATION_END_TIME / 2 - 100;
                                assert_approx_equals(player.currentTime, expectedTime, EPSILON,
                                    'After player is unpaused by AnimationPlayer.play() ' +
                                    'current time should go on decreasing');
                                t.done();
                            });
                        }, 50);
                    });
                }, 50);
            });
        }, 50);
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'Test that after player is unpaused by AnimationPlayer.play() ' +
    'current time goes on decreasing if playback rate < 0');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(t));
    var times = [-1, player.source.computedTiming.endTime,
                 player.source.computedTiming.endTime + 1];

    player.ready.then(t.step_func(function() {
        times.forEach(function(time) {
            player.currentTime = time;
            player.play();

            assert_equals(player.currentTime, 0, 'AnimationPlayer.play() call should seek ' +
                'current time to zero for time ' + time);
        });
        t.done();
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'AnimationPlayer.play() seeks current time to zero if current time < 0 or >= ' +
    'source content end');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(t));
    player.playbackRate = -1;
    var times = [-1, 0, player.source.computedTiming.endTime + 1];

    player.ready.then(t.step_func(function() {
        times.forEach(function(time) {
            player.currentTime = time;
            player.play();

            assert_equals(player.currentTime, player.source.computedTiming.endTime,
                'AnimationPlayer.play() call should seek current time to source content end ' +
                'for time ' + time);
        });
        t.done();
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'AnimationPlayer.play() seeks current time to source content end if current time <= 0 or > ' +
    'source content end and playback rate < 0');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(t));

    player.ready.then(t.step_func(function() {
        times.forEach(function(time) {
            player.currentTime = ANIMATION_END_TIME / 2;
            player.play();

            assert_equals(player.currentTime, ANIMATION_END_TIME / 2,
                'AnimationPlayer.play() call should not change current time');
        });
        t.done();
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'AnimationPlayer.play() does not change current time current time >= 0 and < ' +
    'source content end and playback rate > 0');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(t));
    player.playbackRate = -1;

    player.ready.then(t.step_func(function() {
        times.forEach(function(time) {
            player.currentTime = ANIMATION_END_TIME / 2;
            player.play();

            assert_equals(player.currentTime, ANIMATION_END_TIME / 2,
                'AnimationPlayer.play() call should not change current time');
        });
        t.done();
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'AnimationPlayer.play() does not change current time current time > 0 and <= ' +
    'source content end and playback rate < 0');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));
    player.currentTime = player.source.computedTiming.endTime + 100;

    player.ready.then(t.step_func(function() {
        // wait for 50 ms
        setTimeout(function() {
            t.step(function () {
                assert_equals(player.currentTime, player.source.computedTiming.endTime + 100,
                    'currentTime value should not be changed');
                player.play();
                // wait for 50 ms
                setTimeout(function() {
                    t.step(function() {
                        assert_approx_equals(player.currentTime, 50, EPSILON,
                            'After AnimationPlayer.play() call ' +
                            'current time should go on increasing');
                        t.done();
                    });
                }, 50);
            });
        }, 50);
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'Test that if AnimationPlayer.play() is called when current time > source end time ' +
    'seek time is set to zero and starts increasing');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));
    player.playbackRate = -1;
    player.currentTime = -10;

    player.ready.then(t.step_func(function() {
        // wait for 50 ms
        setTimeout(function() {
            t.step(function () {
                assert_equals(player.currentTime, -10, 'currentTime value should not be changed');
                player.play();
                // wait for 50 ms
                setTimeout(function() {
                    t.step(function() {
                        assert_approx_equals(player.currentTime,
                            player.source.computedTiming.endTime - 50, EPSILON,
                            'After AnimationPlayer.play() call' +
                            'current time should go on decreasing');
                        t.done();
                    });
                }, 50);
            });
        }, 50);
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'Test that if AnimationPlayer.play() is called when current time < 0 ' +
    'seek time is set to source end time and starts decreasing if playback rate < 0');
</script>
</body>
